+++ cdb10/1.0/site/apidocs/allclasses-noframe.html
@@ -14,7 +14,7 @@
 <div class="indexContainer">
 <ul>
 <li><a href="org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure">Capability1Tests</a></li>
-<li><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2">Capability2Tests</a></li>
+<li><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></li>
 <li><a href="org/opengis/cite/cdb10/util/ClientUtils.html" title="class in org.opengis.cite.cdb10.util">ClientUtils</a></li>
 <li><a href="org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a></li>
 <li><a href="org/opengis/cite/cdb10/ErrorMessage.html" title="class in org.opengis.cite.cdb10">ErrorMessage</a></li>
