+++ cdb10/1.0/site/apidocs/index-all.html
@@ -154,17 +154,17 @@
 </dd>
 <dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html#Capability1Tests--">Capability1Tests()</a></span> - Constructor for class org.opengis.cite.cdb10.CDBStructure.<a href="org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure">Capability1Tests</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2"><span class="typeNameLink">Capability2Tests</span></a> - Class in <a href="org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.level2</a></dt>
+<dt><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning"><span class="typeNameLink">Capability2Tests</span></a> - Class in <a href="org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></dt>
 <dd>
 <div class="block">Includes various tests of capability 2.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#Capability2Tests--">Capability2Tests()</a></span> - Constructor for class org.opengis.cite.cdb10.level2.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2">Capability2Tests</a></dt>
+<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#Capability2Tests--">Capability2Tests()</a></span> - Constructor for class org.opengis.cite.cdb10.MetadataAndVersioning.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#checkLength--">checkLength()</a></span> - Method in class org.opengis.cite.cdb10.level2.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2">Capability2Tests</a></dt>
+<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#checkLength--">checkLength()</a></span> - Method in class org.opengis.cite.cdb10.MetadataAndVersioning.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></dt>
 <dd>
 <div class="block">Checks the result of the length function.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#checkPreconditions--">checkPreconditions()</a></span> - Method in class org.opengis.cite.cdb10.level2.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2">Capability2Tests</a></dt>
+<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#checkPreconditions--">checkPreconditions()</a></span> - Method in class org.opengis.cite.cdb10.MetadataAndVersioning.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></dt>
 <dd>
 <div class="block">Run conformance level 2 tests only if the preconditions are satisfied.</div>
 </dd>
@@ -181,7 +181,7 @@
 </dd>
 <dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/util/ClientUtils.html#ClientUtils--">ClientUtils()</a></span> - Constructor for class org.opengis.cite.cdb10.util.<a href="org/opengis/cite/cdb10/util/ClientUtils.html" title="class in org.opengis.cite.cdb10.util">ClientUtils</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#codePoint--">codePoint()</a></span> - Method in class org.opengis.cite.cdb10.level2.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2">Capability2Tests</a></dt>
+<dt><span class="memberNameLink"><a href="org/opengis/cite/cdb10/level2/Capability2Tests.html#codePoint--">codePoint()</a></span> - Method in class org.opengis.cite.cdb10.MetadataAndVersioning.<a href="org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></dt>
 <dd>
 <div class="block">Checks the Unicode code point value of the first character.</div>
 </dd>
@@ -468,7 +468,7 @@
 <dd>
 <div class="block">Conformance Level 1 includes basic facilities.</div>
 </dd>
-<dt><a href="org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.level2</a> - package org.opengis.cite.cdb10.level2</dt>
+<dt><a href="org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a> - package org.opengis.cite.cdb10.MetadataAndVersioning</dt>
 <dd>
 <div class="block">Conformance Level 2 extends Level 1.</div>
 </dd>
