+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/CommonFixture.html
@@ -110,7 +110,7 @@ var activeTableTab = "activeTableTab";
 <li class="blockList">
 <dl>
 <dt>Direct Known Subclasses:</dt>
-<dd><a href="../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure">Capability1Tests</a>, <a href="../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2">Capability2Tests</a></dd>
+<dd><a href="../../../../org/opengis/cite/cdb10/CDBStructure/Capability1Tests.html" title="class in org.opengis.cite.cdb10.CDBStructure">Capability1Tests</a>, <a href="../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></dd>
 </dl>
 <hr>
 <br>
