+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/class-use/CommonFixture.html
@@ -90,7 +90,7 @@
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="#org.opengis.cite.cdb10.level2">org.opengis.cite.cdb10.level2</a></td>
+<td class="colFirst"><a href="#org.opengis.cite.cdb10.MetadataAndVersioning">org.opengis.cite.cdb10.MetadataAndVersioning</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 2 extends Level 1.</div>
 </td>
@@ -120,12 +120,12 @@
 </tbody>
 </table>
 </li>
-<li class="blockList"><a name="org.opengis.cite.cdb10.level2">
+<li class="blockList"><a name="org.opengis.cite.cdb10.MetadataAndVersioning">
 <!--   -->
 </a>
-<h3>Uses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.level2</a></h3>
+<h3>Uses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></h3>
 <table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing subclasses, and an explanation">
-<caption><span>Subclasses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.level2</a></span><span class="tabEnd">&nbsp;</span></caption>
+<caption><span>Subclasses of <a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">CommonFixture</a> in <a href="../../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></span><span class="tabEnd">&nbsp;</span></caption>
 <tr>
 <th class="colFirst" scope="col">Modifier and Type</th>
 <th class="colLast" scope="col">Class and Description</th>
@@ -133,7 +133,7 @@
 <tbody>
 <tr class="altColor">
 <td class="colFirst"><code>class&nbsp;</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.level2">Capability2Tests</a></span></code>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../org/opengis/cite/cdb10/level2/Capability2Tests.html" title="class in org.opengis.cite.cdb10.MetadataAndVersioning">Capability2Tests</a></span></code>
 <div class="block">Includes various tests of capability 2.</div>
 </td>
 </tr>
