+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/level2/Capability2Tests.html
@@ -93,7 +93,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">org.opengis.cite.cdb10.level2</div>
+<div class="subTitle">org.opengis.cite.cdb10.MetadataAndVersioning</div>
 <h2 title="Class Capability2Tests" class="title">Class Capability2Tests</h2>
 </div>
 <div class="contentContainer">
@@ -104,7 +104,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="../../../../../org/opengis/cite/cdb10/CommonFixture.html" title="class in org.opengis.cite.cdb10">org.opengis.cite.cdb10.CommonFixture</a></li>
 <li>
 <ul class="inheritance">
-<li>org.opengis.cite.cdb10.level2.Capability2Tests</li>
+<li>org.opengis.cite.cdb10.MetadataAndVersioning.Capability2Tests</li>
 </ul>
 </li>
 </ul>
