+++ cdb10/1.0/site/apidocs/org/opengis/cite/cdb10/package-use.html
@@ -97,7 +97,7 @@
 </td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a href="#org.opengis.cite.cdb10.level2">org.opengis.cite.cdb10.level2</a></td>
+<td class="colFirst"><a href="#org.opengis.cite.cdb10.MetadataAndVersioning">org.opengis.cite.cdb10.MetadataAndVersioning</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 2 extends Level 1.</div>
 </td>
@@ -145,17 +145,17 @@
 </tbody>
 </table>
 </li>
-<li class="blockList"><a name="org.opengis.cite.cdb10.level2">
+<li class="blockList"><a name="org.opengis.cite.cdb10.MetadataAndVersioning">
 <!--   -->
 </a>
 <table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing classes, and an explanation">
-<caption><span>Classes in <a href="../../../../org/opengis/cite/cdb10/package-summary.html">org.opengis.cite.cdb10</a> used by <a href="../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.level2</a></span><span class="tabEnd">&nbsp;</span></caption>
+<caption><span>Classes in <a href="../../../../org/opengis/cite/cdb10/package-summary.html">org.opengis.cite.cdb10</a> used by <a href="../../../../org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></span><span class="tabEnd">&nbsp;</span></caption>
 <tr>
 <th class="colOne" scope="col">Class and Description</th>
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colOne"><a href="../../../../org/opengis/cite/cdb10/class-use/CommonFixture.html#org.opengis.cite.cdb10.level2">CommonFixture</a>
+<td class="colOne"><a href="../../../../org/opengis/cite/cdb10/class-use/CommonFixture.html#org.opengis.cite.cdb10.MetadataAndVersioning">CommonFixture</a>
 <div class="block">A supporting base class that sets up a common test fixture.</div>
 </td>
 </tr>
