+++ cdb10/1.0/site/apidocs/overview-frame.html
@@ -16,7 +16,7 @@
 <ul title="Packages">
 <li><a href="org/opengis/cite/cdb10/package-frame.html" target="packageFrame">org.opengis.cite.cdb10</a></li>
 <li><a href="org/opengis/cite/cdb10/CDBStructure/package-frame.html" target="packageFrame">org.opengis.cite.cdb10.CDBStructure</a></li>
-<li><a href="org/opengis/cite/cdb10/level2/package-frame.html" target="packageFrame">org.opengis.cite.cdb10.level2</a></li>
+<li><a href="org/opengis/cite/cdb10/level2/package-frame.html" target="packageFrame">org.opengis.cite.cdb10.MetadataAndVersioning</a></li>
 <li><a href="org/opengis/cite/cdb10/util/package-frame.html" target="packageFrame">org.opengis.cite.cdb10.util</a></li>
 </ul>
 </div>
