+++ cdb10/1.0/site/apidocs/overview-summary.html
@@ -101,7 +101,7 @@
 </td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a href="org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.level2</a></td>
+<td class="colFirst"><a href="org/opengis/cite/cdb10/level2/package-summary.html">org.opengis.cite.cdb10.MetadataAndVersioning</a></td>
 <td class="colLast">
 <div class="block">Conformance Level 2 extends Level 1.</div>
 </td>
