+++ src/main/resources/org/opengis/cite/cdb10/testng.xml
@@ -14,9 +14,9 @@
       <package name="org.opengis.cite.cdb10.CDBStructure" />
     </packages>
   </test>
- <!-- <test name="Conformance Level 2">
+  <test name="Metadata And Versioning">
     <packages>
-      <package name="org.opengis.cite.cdb10.level2" />
+      <package name="org.opengis.cite.cdb10.MetadataAndVersioning" />
     </packages>
-  </test> -->
+  </test>
 </suite>
